refactor: main의 책임 분리
